flake8 fixes. refactor benchmark utility.
