+++ .flake8
@@ -10,5 +10,4 @@ exclude =
     scent.py
 per-file-ignores = 
     NiaPy/tests/*:B007,D100,D101,D102,D107,D208,D211,D413,E117,E302,E701,E704,F841,S110,W191,W293,D204,RST210
-    NiaPy/algorithms/*:B305,B007,D100,D101,D104,D102,D107,D204,D206,D413,E101,E117,E126,E302,E305,E701,E704,E711,E741,F812,RST201,RST203,RST204,RST210,RST301,W191
-    NiaPy/util/*:D100,D101,D104,D107,D204,D413,E117,E242,E302,E305,E701,E704,F104,RST201,RST203,RST210,W191
+    NiaPy/algorithms/*:B305,B007,D100,D101,D102,D107,D204,D206,D413,E101,E117,E126,E302,E305,E701,E704,E711,E741,F812,RST201,RST203,RST204,RST210,RST301,W191
